changes to contrast_streching & noise_types files
